--- app/models/prueba.csv
@@ -1,3 +1,0 @@
-name,department,birthday month
-John Smith,Accounting,November
-Erica Meyers,IT,March
