--- app/requirements.txt
@@ -1,156 +1,0 @@
-# This file may be used to create an environment using:
-# $ conda create --name <env> --file <this file>
-# platform: osx-64
-aom=3.5.0=hf0c8a7f_0
-appnope=0.1.2=py310hecd8cb5_1001
-asttokens=2.0.5=pyhd3eb1b0_0
-backcall=0.2.0=pyhd3eb1b0_0
-blas=1.0=mkl
-bottleneck=1.3.4=py310h4e76f89_0
-brotli=1.0.9=hb1e8313_2
-brotlipy=0.7.0=py310hca72f7f_1002
-bzip2=1.0.8=h1de35cc_0
-ca-certificates=2022.12.7=h033912b_0
-cairo=1.16.0=h904041c_1014
-certifi=2022.12.7=pyhd8ed1ab_0
-cffi=1.15.1=py310hc55c11b_0
-charset-normalizer=2.0.4=pyhd3eb1b0_0
-click=8.0.4=py310hecd8cb5_0
-cryptography=38.0.1=py310hf6deb26_0
-cycler=0.11.0=pyhd3eb1b0_0
-dataclasses=0.8=pyh6d0b6a4_7
-dav1d=1.0.0=hac89ed1_1
-debugpy=1.5.1=py310he9d5cce_0
-decorator=5.1.1=pyhd3eb1b0_0
-docutils=0.19=py310h2ec42d9_1
-entrypoints=0.4=py310hecd8cb5_0
-executing=0.8.3=pyhd3eb1b0_0
-expat=2.5.0=hf0c8a7f_0
-flask=2.1.3=py310hecd8cb5_0
-fluidsynth=2.3.0=h07f9e24_0
-font-ttf-dejavu-sans-mono=2.37=hab24e00_0
-font-ttf-inconsolata=3.000=h77eed37_0
-font-ttf-source-code-pro=2.038=h77eed37_0
-font-ttf-ubuntu=0.83=hab24e00_0
-fontconfig=2.14.1=h5bb23bf_0
-fonts-conda-ecosystem=1=0
-fonts-conda-forge=1=0
-fonttools=4.25.0=pyhd3eb1b0_0
-fpdf=1.7.2=pyhd8ed1ab_0
-freetype=2.12.1=h3f81eb7_1
-gettext=0.21.1=h8a4c099_0
-giflib=5.2.1=haf1e3a3_0
-glib=2.74.1=hbc0c0cd_1
-glib-tools=2.74.1=hbc0c0cd_1
-graphite2=1.3.13=h2e338ed_1001
-gst-plugins-base=1.20.3=hda0ba4b_0
-gstreamer=1.21.2=h1d18e73_0
-harfbuzz=5.3.0=h08f8713_0
-icu=70.1=h96cf925_0
-idna=3.4=py310hecd8cb5_0
-intel-openmp=2021.4.0=hecd8cb5_3538
-ipykernel=6.9.1=py310hecd8cb5_0
-ipython=8.3.0=py310hecd8cb5_0
-itsdangerous=2.0.1=pyhd3eb1b0_0
-jedi=0.18.1=py310hecd8cb5_1
-jinja2=3.1.2=py310hecd8cb5_0
-joblib=1.1.0=pyhd3eb1b0_0
-jpeg=9e=hca72f7f_0
-jupyter_client=7.2.2=py310hecd8cb5_0
-jupyter_core=4.10.0=py310hecd8cb5_0
-keras=2.11.0=pyhd8ed1ab_0
-kivy=2.1.0=py310h892de82_2
-kiwisolver=1.4.2=py310he9d5cce_0
-lame=3.100=hb7f2c08_1003
-lcms2=2.12=hf1fd2bf_0
-lerc=4.0.0=hb486fe8_0
-libavif=0.11.1=he9a59b1_0
-libcxx=14.0.6=hccf4f1f_0
-libdeflate=1.14=hb7f2c08_0
-libffi=3.4.2=h0d85af4_5
-libflac=1.4.2=hf0c8a7f_0
-libgfortran=3.0.1=h93005f0_2
-libglib=2.74.1=h4c723e1_1
-libiconv=1.17=hac89ed1_0
-libmad=0.15.1b=h1de35cc_1000
-libogg=1.3.4=h35c211d_1
-libopus=1.3.1=hc929b4f_1
-libpng=1.6.39=ha978bb4_0
-libsndfile=1.1.0=hf0c8a7f_1
-libsodium=1.0.18=h1de35cc_0
-libsqlite=3.40.0=ha978bb4_0
-libtiff=4.4.0=hdb44e8a_4
-libvorbis=1.3.7=h046ec9c_0
-libwebp=1.2.4=hfa4350a_0
-libwebp-base=1.2.4=h775f41a_0
-libzlib=1.2.13=hfd90126_4
-llvm-openmp=12.0.0=h0dcd299_1
-lz4-c=1.9.3=h23ab428_1
-markupsafe=2.1.1=py310hca72f7f_0
-matplotlib=3.5.1=py310hecd8cb5_1
-matplotlib-base=3.5.1=py310hfb0c5b7_1
-matplotlib-inline=0.1.2=pyhd3eb1b0_2
-mkl=2021.4.0=hecd8cb5_637
-mkl-service=2.4.0=py310hca72f7f_0
-mkl_fft=1.3.1=py310hf879493_0
-mkl_random=1.2.2=py310hc081a56_0
-mpg123=1.31.1=hf0c8a7f_0
-munkres=1.1.4=py_0
-ncurses=6.3=hca72f7f_2
-nest-asyncio=1.5.5=py310hecd8cb5_0
-numexpr=2.8.1=py310hdcd3fac_2
-numpy=1.22.3=py310hdcd3fac_0
-numpy-base=1.22.3=py310hfd2de13_0
-openssl=3.0.7=hfd90126_1
-opusfile=0.12=h9cee658_2
-packaging=21.3=pyhd3eb1b0_0
-pandas=1.4.2=py310he9d5cce_0
-parso=0.8.3=pyhd3eb1b0_0
-pcre2=10.40=h1c4e4bc_0
-pexpect=4.8.0=pyhd3eb1b0_3
-pickleshare=0.7.5=pyhd3eb1b0_1003
-pillow=9.0.1=py310hde71d04_0
-pip=21.2.4=py310hecd8cb5_0
-pixman=0.40.0=hbcb3906_0
-plotly-orca=1.3.1=1
-portaudio=19.6.0=hf0c8a7f_7
-prompt-toolkit=3.0.20=pyhd3eb1b0_0
-psutil=5.9.0=py310hca72f7f_0
-ptyprocess=0.7.0=pyhd3eb1b0_2
-pure_eval=0.2.2=pyhd3eb1b0_0
-pycparser=2.21=pyhd3eb1b0_0
-pygments=2.11.2=pyhd3eb1b0_0
-pyopenssl=22.0.0=pyhd3eb1b0_0
-pyparsing=3.0.4=pyhd3eb1b0_0
-pysocks=1.7.1=py310hecd8cb5_0
-python=3.10.8=he7542f4_0_cpython
-python-dateutil=2.8.2=pyhd3eb1b0_0
-python_abi=3.10=3_cp310
-pytz=2022.1=py310hecd8cb5_0
-pyzmq=22.3.0=py310he9d5cce_2
-readline=8.1.2=hca72f7f_1
-requests=2.28.1=py310hecd8cb5_0
-scikit-learn=1.0.2=py310hc081a56_0
-scipy=1.7.3=py310h3dd3380_0
-sdl2=2.26.1=hf0c8a7f_1
-sdl2_image=2.6.2=h20bcf9f_2
-sdl2_mixer=2.6.2=hf0c8a7f_1
-sdl2_ttf=2.20.1=hf19bd59_0
-setuptools=61.2.0=py310hecd8cb5_0
-six=1.16.0=pyhd3eb1b0_1
-sqlite=3.38.5=h707629a_0
-stack_data=0.2.0=pyhd3eb1b0_0
-threadpoolctl=2.2.0=pyh0d69192_0
-tk=8.6.12=h5d9f67b_0
-tornado=6.1=py310hca72f7f_0
-traitlets=5.1.1=pyhd3eb1b0_0
-tzdata=2022a=hda174b7_0
-urllib3=1.26.12=py310hecd8cb5_0
-wcwidth=0.2.5=pyhd3eb1b0_0
-werkzeug=2.0.3=pyhd3eb1b0_0
-wheel=0.37.1=pyhd3eb1b0_0
-wtforms=2.3.3=pyhd3eb1b0_0
-xz=5.2.6=h775f41a_0
-zeromq=4.3.4=h23ab428_0
-zlib=1.2.13=hfd90126_4
-zstd=1.5.2=hcb37349_0
