--- app/trainer.py
@@ -1,2 +1,0 @@
-def train_model(dataset):
-    print("Hola mundo")
