+++ lens_effect_project/model/const.py
@@ -1,0 +1,18 @@
+BATCH_SIZE_256 = 256
+BATCH_SIZE_128 = 128
+BATCH_SIZE_64 = 64
+BATCH_SIZE_32 = 32
+
+TWO_WORKERS = 2
+FOUR_WORKERS = 4
+SIX_WORKERS = 6
+
+TRAIN = 0.2
+
+FALSE_VALUE = False
+TRUE_VALUE = True
+
+PATH_OF_FIGURES = '/home/ariel/Documents/models-aceron/lens-data/figures'
+PATH_OF_TSV = '/home/ariel/Documents/models-aceron/lens-data/utils/training_lens.tsv'
+
+PATH_OF_PREDOs_PC = '/home/pedroarturo/TF2/Pretrain/imagenes_no_clasificadas_desi'
